app-backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.2
Pillow==10.3.0
django-jazzmin==2.6.0
fcm-django==2.3.1
psycopg2-binary==2.9.9
